+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Print `EventSet` timestamps as datetimes instead of float.
 - Support `sampling` argument in `EventSet.cumsum()` operator.
+- Using utf-8 codec to support non-ascii in string values.
 
 ### Fixes